pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-agent"
version = "0.1.0"
description = "AI-powered hybrid job application agent"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "openai>=1.40",
    "anthropic>=0.34",
    "playwright>=1.45",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "pyyaml>=6.0",
    "typer[all]>=0.12",
    "rich>=13.7",
    "python-dotenv>=1.0",
    "pdfplumber>=0.11",
    "aiosqlite>=0.20",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
]

[project.scripts]
job-agent = "src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
